feat: connect post quote and get feed api
