+++ src/app/(route)/write/page.tsx
@@ -22,6 +22,7 @@ import FlexContainer from '@/app/_components/atoms/FlexContainer';
 import useDebounce from '@/app/_hooks/useDebounce';
 import { useCompletion } from 'ai/react';
 import { AICompletionType } from '@/app/_constant/ai';
+import { StoryType } from '@/app/_constant/story';
 
 const MAIN_TEXT = '타인에게서\n자신의 이야기를\n발견하세요.';
 const INTRO_TEXT = '타인에게서\n자신의 이야기를\n발견하세요.';
@@ -267,7 +268,37 @@ export default function Page() {
               <div className={`${styles.wrap} final ani_leftToRight`}>
                 <div className={styles.text}>{MAIN_TEXT}</div>
                 <div className={styles.vertiline}></div>
-                <div className={styles.btn}>흘려보내기</div>
+                <div
+                  className={styles.btn}
+                  onClick={async () => {
+                    try {
+                      const response = await fetch('/api/story', {
+                        method: 'POST',
+                        body: JSON.stringify({
+                          pen_name:
+                            (story.author?.length ?? 0) > 0
+                              ? story.author
+                              : undefined,
+                          content: story.content,
+                          type:
+                            _type === 'story'
+                              ? StoryType.ESSAY
+                              : StoryType.QUOTE,
+                          title:
+                            (story.title?.length ?? 0) > 0
+                              ? story.title
+                              : undefined,
+                        }),
+                      });
+                      const result = await response.json();
+                      console.log(result);
+                    } catch (error) {
+                      console.error(error);
+                    }
+                  }}
+                >
+                  흘려보내기
+                </div>
               </div>
             </FlexContainer>
           </div>
